User field adjustments
This is missing the field organization adjustments because I am unsure of how this info will be stored.

This uses the install hook to adjust the fields on the user profile.

I'm not sure what to do with the features -- if I need to do anything at all. Advice is most appreciated!