#38: latest slides
Please note: I also made changes to the view -- I added the class .latest-slides for the block and added the session class for the rows. This gives a little breathing room between the items.

You can check out some sample data in the block on Test: http://test-nyccamp2014.gotpantheon.com/, the block is at the bottom.